Refactor User Store
